@@ -1,58 +1,58 @@
 {
-  "name": "frontend",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "auth0-js": "^9.3.0",
-    "axios": "^0.17.1",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "gulp": "^3.9.1",
-    "gulp-changed": "^3.2.0",
-    "gulp-clean-css": "^3.9.2",
-    "gulp-compass": "^2.1.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-sass": "^3.1.0",
-    "gulp-uglify": "^3.0.0",
-    "lodash": "^4.17.5",
-    "rc-checkbox": "^2.1.4",
-    "react": "^16.2.0",
-    "react-accessible-accordion": "^2.2.1",
-    "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.2.0",
-    "react-excel-workbook": "^0.0.4",
-    "react-file-reader": "^1.1.3",
-    "react-input-range": "^1.3.0",
-    "react-joyride": "^1.11.4",
-    "react-month-picker": "^1.3.5",
-    "react-month-picker-input": "^1.0.0",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.1",
-    "react-sticky-table": "^2.0.0",
-    "react-tabs": "^2.2.1",
-    "react-widgets": "^4.1.3",
-    "recharts": "^1.0.0-beta.10",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0",
-    "source-map-explorer": "^1.5.0"
-  },
-  "scripts": {
-    "analyze": "source-map-explorer build/static/js/main.*",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "devDependencies": {
-    "babel-core": "^5.6.18",
-    "babel-loader": "^5.1.4",
-    "babel-plugin-react-transform": "^1.1.0",
-    "express": "^4.13.3",
-    "webpack": "^1.9.11",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.2.0"
-  }
+	"name": "frontend",
+	"version": "0.1.0",
+	"private": true,
+	"dependencies": {
+		"auth0-js": "^9.3.0",
+		"axios": "^0.17.1",
+		"babel-cli": "^6.26.0",
+		"babel-plugin-transform-object-rest-spread": "^6.26.0",
+		"babel-preset-react": "^6.24.1",
+		"babel-preset-stage-0": "^6.24.1",
+		"gulp": "^3.9.1",
+		"gulp-changed": "^3.2.0",
+		"gulp-clean-css": "^3.9.2",
+		"gulp-compass": "^2.1.0",
+		"gulp-rename": "^1.2.2",
+		"gulp-sass": "^3.1.0",
+		"gulp-uglify": "^3.0.0",
+		"lodash": "^4.17.5",
+		"rc-checkbox": "^2.1.4",
+		"react": "^16.3.0",
+		"react-accessible-accordion": "^2.2.1",
+		"react-bootstrap": "^0.32.1",
+		"react-dom": "^16.3.0",
+		"react-excel-workbook": "^0.0.4",
+		"react-file-reader": "^1.1.3",
+		"react-input-range": "^1.3.0",
+		"react-joyride": "^1.11.4",
+		"react-month-picker": "^1.3.5",
+		"react-month-picker-input": "^1.0.0",
+		"react-redux": "^5.0.7",
+		"react-router-dom": "^4.2.2",
+		"react-scripts": "1.1.1",
+		"react-sticky-table": "^2.0.0",
+		"react-tabs": "^2.2.1",
+		"react-widgets": "^4.1.3",
+		"recharts": "^1.0.0-beta.10",
+		"redux": "^3.7.2",
+		"redux-thunk": "^2.2.0",
+		"source-map-explorer": "^1.5.0"
+	},
+	"scripts": {
+		"analyze": "source-map-explorer build/static/js/main.*",
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"test": "react-scripts test --env=jsdom",
+		"eject": "react-scripts eject"
+	},
+	"devDependencies": {
+		"babel-core": "^5.6.18",
+		"babel-loader": "^5.1.4",
+		"babel-plugin-react-transform": "^1.1.0",
+		"express": "^4.13.3",
+		"webpack": "^1.9.11",
+		"webpack-dev-middleware": "^1.2.0",
+		"webpack-hot-middleware": "^2.2.0"
+	}
 }
