@@ -62,7 +62,7 @@ class CreateCalendar extends Component {
 		}, {})
 	}
 
-	componentWillReceiveProps(nextProps) {
+	UNSAFE_componentWillReceiveProps(nextProps) {
 		if (nextProps.ads.length > 0 && !_.isEqual(this.props, nextProps)) {
 			const { year, month } = this.getMinDate(
 				nextProps.ads,
