@@ -8,7 +8,6 @@ import ColorChart from './colorChart'
 import MessagingCode from './messagingCode'
 import GetKPIs from './getKPIs'
 import UserRights from './userRights'
-import Slugify from './slugify'
 
 export {
 	Api,
@@ -20,6 +19,5 @@ export {
 	ColorChart,
 	MessagingCode,
 	GetKPIs,
-	UserRights,
-	Slugify
+	UserRights
 }
